+++ examples/bluepill-receiver/src/bin/main.rs
@@ -17,8 +17,9 @@ use heapless::consts::*;
 use heapless::spsc::Queue;
 
 use infrared::{
-    Receiver, ReceiverState, remote::RemoteControl,
-    nec::*, nec::remotes::SpecialForMp3,
+    Receiver, ReceiverState,
+    nec::*,
+    rc5::*,
     rc6::*,
 };
 
@@ -27,12 +28,15 @@ const FREQ: u32 = 40_000;
 static mut TIMER: Option<Timer<TIM2>> = None;
 static mut IRPIN: Option<PB8<Input<Floating>>> = None;
 
+static mut NEC: Option<NecReceiver> = None;
+static mut NES: Option<NecSamsungReceiver> = None;
+static mut RC5: Option<Rc5Receiver> = None;
 static mut RC6: Option<Rc6Receiver> = None;
-static mut NEC: Option<NecReceiver<u32>> = None;
 
 // Command Queues
+static mut NECQ: Option<Queue<NecCommand, U8>> = None;
+static mut RC5Q: Option<Queue<Rc5Command, U8>> = None;
 static mut RC6Q: Option<Queue<Rc6Command, U8>> = None;
-static mut NECQ: Option<Queue<NecCommand<u32>, U8>> = None;
 
 #[entry]
 fn main() -> ! {
@@ -62,17 +66,21 @@ fn main() -> ! {
         TIMER.replace(timer);
         IRPIN.replace(irpin);
 
-        NEC.replace(NecReceiver::new(NecType::Standard, FREQ));
+        NEC.replace(NecReceiver::new(FREQ));
+        NES.replace(NecSamsungReceiver::new(FREQ));
+        RC5.replace(Rc5Receiver::new(FREQ));
         RC6.replace(Rc6Receiver::new(FREQ));
     }
 
     // Initialize the queues
     unsafe {
         NECQ.replace(Queue::new());
+        RC5Q.replace(Queue::new());
         RC6Q.replace(Queue::new());
     };
 
     let mut necq = unsafe { NECQ.as_mut().unwrap().split().1 };
+    let mut rc5q = unsafe { RC5Q.as_mut().unwrap().split().1 };
     let mut rc6q = unsafe { RC6Q.as_mut().unwrap().split().1 };
 
     // Enable the timer interrupt
@@ -81,14 +89,16 @@ fn main() -> ! {
     hprintln!("Ready!").unwrap();
 
     loop {
+        if let Some(cmd) = necq.dequeue() {
+            hprintln!("nec: {:?}", cmd).unwrap();
+        }
 
-        if let Some(NecCommand::Payload(cmd)) = necq.dequeue() {
-            let cmd = SpecialForMp3.decode(cmd);
-            hprintln!("{:?}", cmd).unwrap();
+        if let Some(cmd) = rc5q.dequeue() {
+            hprintln!("rc5: {:?}", cmd).unwrap();
         }
 
         if let Some(cmd) = rc6q.dequeue() {
-            hprintln!("{:?}", cmd).unwrap();
+            hprintln!("rc6: {:?}", cmd).unwrap();
         }
     }
 }
@@ -109,32 +119,54 @@ fn TIM2() {
         let rising = new_pinval;
 
         let mut necq = unsafe { NECQ.as_mut().unwrap().split().0 };
+        let mut rc5q = unsafe { RC5Q.as_mut().unwrap().split().0 };
         let mut rc6q = unsafe { RC6Q.as_mut().unwrap().split().0 };
 
         let nec = unsafe { NEC.as_mut().unwrap() };
+        let nes = unsafe { NES.as_mut().unwrap() };
+        let rc5 = unsafe { RC5.as_mut().unwrap() };
         let rc6 = unsafe { RC6.as_mut().unwrap() };
 
         // NEC
-        match nec.event(rising, *COUNT) {
+        match nec.sample_edge(rising, *COUNT) {
             ReceiverState::Done(cmd) => {
                 necq.enqueue(cmd).ok().unwrap();
                 nec.reset();
             },
-            ReceiverState::Err(_err) => nec.reset(),
+            ReceiverState::Error(_err) => nec.reset(),
+            _ => (),
+        }
+
+        // Nec with Samsung flavour
+        match nes.sample_edge(rising, *COUNT) {
+            ReceiverState::Done(cmd) => {
+                necq.enqueue(cmd).unwrap();
+                nes.reset();
+            },
+            ReceiverState::Error(_err) => nes.reset(),
+            _ => (),
+        }
+
+        match rc5.sample_edge(rising, *COUNT) {
+            ReceiverState::Done(cmd) => {
+                rc5q.enqueue(cmd).unwrap();
+                rc5.reset();
+            },
+            ReceiverState::Error(_err) => rc5.reset(),
             _ => (),
         }
 
         // RC6
-        match rc6.event(rising, *COUNT) {
+        match rc6.sample_edge(rising, *COUNT) {
             ReceiverState::Done(cmd) => {
-                rc6q.enqueue(cmd).ok().unwrap();
+                rc6q.enqueue(cmd).unwrap();
                 rc6.reset();
             },
-            ReceiverState::Err(_err) => rc6.reset(),
+            ReceiverState::Error(_err) => rc6.reset(),
             _ => (),
         }
     }
 
     *PINVAL = new_pinval;
-    *COUNT += 1;
+    *COUNT = COUNT.wrapping_add(1);
 }