+++ src/protocols/mod.rs
@@ -1,3 +1,7 @@
+#[cfg(feature="nec")]
 pub mod nec;
-pub mod rc6;
+#[cfg(feature="rc5")]
 pub mod rc5;
+#[cfg(feature="rc6")]
+pub mod rc6;
+