+++ src/protocols/nec/mod.rs
@@ -3,13 +3,49 @@ pub mod remotes;
 pub mod receiver;
 pub mod transmitter;
 
-pub use receiver::{NecCommand, NecError, NecReceiver, NecResult};
-pub use transmitter::NecTransmitter;
+pub use receiver::{NecCommand, NecError, NecTypeReceiver, NecResult};
+pub use transmitter::NecTypeTransmitter;
 
-pub enum NecType {
-    Standard,
-    Samsung,
+pub type NecReceiver = NecTypeReceiver<StandardType>;
+pub type NecSamsungReceiver = NecTypeReceiver<SamsungType>;
+
+pub type NecTransmitter = NecTypeTransmitter<SamsungType>;
+
+pub struct SamsungType;
+pub struct StandardType;
+
+pub trait NecTypeTrait {
+    const TIMING: Timing;
+    const ADDR_BITS: usize;
+    const CMD_BITS: usize;
 }
+
+impl NecTypeTrait for SamsungType {
+    const TIMING: Timing = Timing {
+        header_high: 4500,
+        header_low: 4500,
+        repeat_low: 2250,
+        zero_low: 560,
+        data_high: 560,
+        one_low: 1690,
+    };
+    const ADDR_BITS: usize = 8;
+    const CMD_BITS: usize = 8;
+}
+
+impl NecTypeTrait for StandardType {
+    const TIMING: Timing = Timing {
+        header_high: 9000,
+        header_low: 4500,
+        repeat_low: 2250,
+        data_high: 560,
+        zero_low: 560,
+        one_low: 1690,
+    };
+    const ADDR_BITS: usize = 8;
+    const CMD_BITS: usize = 8;
+}
+
 
 pub struct Timing {
     header_high: u32,
@@ -20,20 +56,18 @@ pub struct Timing {
     one_low: u32,
 }
 
-const STANDARD_TIMING: Timing = Timing {
-    header_high: 9000,
-    header_low: 4500,
-    repeat_low: 2250,
-    data_high: 560,
-    zero_low: 560,
-    one_low: 1690,
-};
+#[cfg(test)]
+mod tests {
+    use crate::protocols::nec::NecReceiver;
 
-const SAMSUNG_TIMING: Timing = Timing {
-    header_high: 4500,
-    header_low: 4500,
-    repeat_low: 2250,
-    zero_low: 560,
-    data_high: 560,
-    one_low: 1690,
-};
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+
+        let _recv = NecReceiver::new(40_000);
+
+
+
+
+    }
+}