+++ src/remote.rs
@@ -1,7 +1,7 @@
-pub trait RemoteControl {
+pub trait RemoteControl<CMD> {
     type Action;
 
-    fn decode(&self, raw: u32) -> Option<Self::Action>;
+    fn decode(&self, raw: CMD) -> Option<Self::Action>;
 
     fn encode(&self, cmd: Self::Action) -> u32;
 }